refactor: extract statistics calculation to useHeritageStats hook (Phase 2 - part 5)
- Create useHeritageStats hook (64 lines) for heritage statistics calculations
- Update StatsDashboard.tsx to use the hook (601 → 550 lines)
- Extract complex age parsing and site categorization logic
- Statistics now reusable across application

Benefits:
- Separated calculation logic from presentation
- Hook can be used in other components
- Easier to test statistics logic independently
- Reduced StatsDashboard complexity

All 184 tests passing
Linter clean (0 errors, 0 warnings)
